Update default.css
This is an alternative stylesheet - moving to responsive (rem based), clean and minimal. 
It includes a @print section that on my  firefox on win7 works only if loaded as external CSS, not when embedded.
Still work in progress (some trouble with images) but hopefully it will work with my proposed modification to html exporter.
It was hard to move in github for my first time, I'm not a programmer, just a graphic hobbist, so I would appreciate a feedback.